src/sentiment/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
groq==0.4.2
python-dateutil==2.8.2
lxml==5.1.0
